requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
twilio==8.11.0
sarvamai>=0.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
